+++ src/launch/github/repo.py
@@ -19,21 +19,6 @@ def get_github_repos(
     return repos
 
 
-def clone_repository(repository_url: str, target: str, branch: str) -> Repo:
-    try:
-        logger.info(f"Attempting to clone repository: {repository_url} into {target}")
-        repository = Repo.clone_from(repository_url, target, branch=branch)
-        logger.info(f"Repository {repository_url} cloned successfully to {target}")
-    except git.GitCommandError as e:
-        logger.error(
-            f"Error occurred while cloning the repository: {repository_url}, Error: {e}"
-        )
-        raise RuntimeError(
-            f"An error occurred while cloning the repository: {repository_url}"
-        ) from e
-    return repository
-
-
 def create_repository(
     g: Github,
     organization: str,
